Правки, правки, исправления!

+++ src-board/STM/src/ringbuf_varsize.c
@@ -106,7 +106,7 @@ void * rscs_ringbuf_varsize_see_from_head(rscs_ringbuf_varsize_t * buf, size_t s
 void * rscs_ringbuf_varsize_see_from_tail(rscs_ringbuf_varsize_t * buf, size_t shift) {
 	if((shift * buf->element_size) >= buf->size) return NULL;
 
-	int i = buf->tail + (shift * buf->element_size);
+	size_t i = buf->tail + (shift * buf->element_size);
 	if(i >= buf->fullsize) i -= buf->fullsize;
 
 	return buf->buffer + i;
